Created Pagination component. Passing necessary props, reducing PartTable code
